Solucion al conectar con la bd

+++ app/src/main/java/com/example/miniproyecto1/data/ChallengeDBFirebase.kt
@@ -17,7 +17,12 @@ class ChallengeDBFirebase @Inject constructor(
     suspend fun saveChallenge(challenge: Challenge): String? {
         return try {
             val documentRef = challengeCollection.add(challenge).await()
+            challenge.id = documentRef.id
+            print(documentRef)
+            challengeCollection.document(documentRef.id).set(challenge).await()
             documentRef.id
+
+
         } catch (e: Exception) {
             e.printStackTrace()
             null
@@ -37,6 +42,10 @@ class ChallengeDBFirebase @Inject constructor(
     }
 
     suspend fun deleteChallenge(documentId: String): Boolean {
+        if (documentId.isEmpty()) {
+            println("Error: El documentId no puede estar vacío.")
+            return false
+        }
         return try {
             challengeCollection.document(documentId).delete().await()
             true
@@ -48,6 +57,7 @@ class ChallengeDBFirebase @Inject constructor(
 
     suspend fun updateChallenge(challenge: Challenge): Boolean {
         return try {
+            print("updateChallenge firebase"+challenge.id)
             challengeCollection.document(challenge.id).set(challenge).await()
             true
         } catch (e: Exception) {
